# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SFMC Development Suite API
  description: |
    The Syngenta SFMC Development Suite API provides AI-powered development tools for Salesforce Marketing Cloud developers.
    
    ## Features
    - AI-powered code generation for SQL, AMPScript, SSJS, CSS, and HTML
    - Advanced debugging and code analysis
    - Cloud pages generation with responsive design
    - SFMC integration with OAuth 2.0 authentication
    - Session management and caching
    
    ## Authentication
    Most endpoints require authentication via session tokens. Use the `/api/session` endpoint to establish a session.
    
    ## Rate Limiting
    API calls are rate-limited to prevent abuse. Current limits:
    - 100 requests per minute for authenticated users
    - 20 requests per minute for unauthenticated users
    
  version: 1.0.0
  contact:
    name: SFMC Development Suite Support
    email: support@syngenta.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://sfmc-dev-suite-staging.syngenta.com
    description: Staging server
  - url: https://sfmc-dev-suite.syngenta.com
    description: Production server
paths:
...
  /api/sfmc/authenticate:
    post:
      summary: Authenticate with SFMC
      description: |
        Establish OAuth 2.0 connection with Salesforce Marketing Cloud.
        Returns access and refresh tokens for subsequent API calls.
      tags:
        - SFMC Integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SFMCAuthRequest'
            examples:
              auth_example:
                summary: SFMC authentication
                value:
                  clientId: "your-client-id"
                  clientSecret: "your-client-secret"
                  subdomain: "your-subdomain"
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SFMCAuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/sfmc/data-extensions:
    get:
      summary: Get SFMC data extensions
      description: |
        Retrieve list of data extensions from connected SFMC instance.
        Requires valid SFMC authentication.
      tags:
        - SFMC Integration
      parameters:
        - name: limit
          in: query
          description: Maximum number of data extensions to return
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          description: Number of data extensions to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Data extensions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataExtensionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/sfmc/deploy:
    post:
      summary: Deploy cloud page to SFMC
      description: |
        Deploy generated cloud page to connected SFMC instance.
        Requires valid SFMC authentication and proper permissions.
      tags:
        - SFMC Integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentRequest'
      responses:
        '200':
          description: Deployment successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/session:
    post:
      summary: Create user session
      description: |
        Create a new user session for maintaining state across requests.
        Sessions include conversation history and user preferences.
      tags:
        - Session Management
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      summary: Get current session
      description: Get current session information including preferences and history.
      tags:
        - Session Management
      responses:
        '200':
          description: Session information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          description: Session not found
...
components:
  schemas:
...
    SFMCAuthRequest:
      type: object
      required:
        - clientId
        - clientSecret
        - subdomain
      properties:
        clientId:
          type: string
          description: SFMC application client ID
        clientSecret:
          type: string
          description: SFMC application client secret
        subdomain:
          type: string
          description: SFMC instance subdomain
    SFMCAuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: OAuth access token
        refreshToken:
          type: string
          description: OAuth refresh token
        expiresIn:
          type: integer
          description: Token expiration time in seconds
        tokenType:
          type: string
          example: "Bearer"
    DataExtensionResponse:
      type: object
      properties:
        dataExtensions:
          type: array
          items:
            $ref: '#/components/schemas/DataExtension'
        totalCount:
          type: integer
          description: Total number of data extensions
    DeploymentRequest:
      type: object
      required:
        - page
      properties:
        page:
          $ref: '#/components/schemas/GeneratedPage'
        deploymentOptions:
          $ref: '#/components/schemas/DeploymentOptions'
    DeploymentResponse:
      type: object
      properties:
        deploymentId:
          type: string
        status:
          type: string
          enum: [success, failed, pending]
        url:
          type: string
          description: URL of deployed page
        errors:
          type: array
          items:
            type: string

    # Supporting schemas
    CodeLanguage:
      type: string
      enum: [sql, ampscript, ssjs, css, html, javascript]
...
    Message:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        image:
          type: string
        timestamp:
          type: integer
        metadata:
          type: object
...
    GeneratedPage:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        type:
          type: string
        metadata:
          type: object
...
    DataExtension:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        fields:
          type: array
          items:
            type: object
...
    DeploymentOptions:
      type: object
      properties:
        businessUnit:
          type: string
        folder:
          type: string
        overwrite:
          type: boolean
    SessionCreateRequest:
      type: object
      properties:
        preferences:
          $ref: '#/components/schemas/UserPreferences'
    SessionResponse:
      type: object
      properties:
        sessionId:
          type: string
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        conversationHistory:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    UserPreferences:
      type: object
      properties:
        defaultLanguage:
          $ref: '#/components/schemas/CodeLanguage'
        debugLevel:
          type: string
          enum: [basic, advanced, comprehensive]
        theme:
          type: string
          enum: [light, dark, system]
        autoSave:
          type: boolean
        codeFormatting:
          type: object
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "AUTHENTICATION_ERROR"
            message: "Valid authentication required"
            code: "401"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "INTERNAL_SERVER_ERROR"
            message: "An unexpected error occurred"
            code: "500"
...